New version, tarball. Bug fix for GLX/OpenGL output. There was an unhandled case with core profile data being null, which in turn triggered a bash oddity, where if the IFS is \n for an array, and if the value of one element is '', then bash ignores that and does not simply set an empty array key as you'd expect. The correction was to change the IFS to ^, which worked fine for empty array values.
However, since this bug will impact anyone with empty opengl core profile data, I recommend
updating inxi.

Also, added support for two smaller wm, Sawfish and Afterstep.

This is a good source for lists of wm:  http://www.xwinman.org/ http://www.xwinman.org/others.php

However, that does not show how to ID it, so i have to do it on a case by case, but I'll
add an issue for showing how to get your wm of choice if it's missing to inxi.